流程处理上下文

+++ src/main/java/com/wflow/service/impl/FormGroupServiceImpl.java
@@ -4,10 +4,12 @@ import cn.hutool.core.collection.CollectionUtil;
 import com.alibaba.fastjson2.JSONArray;
 import com.alibaba.fastjson2.JSONObject;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
+import com.wflow.bean.FlowProcessContext;
 import com.wflow.bean.entity.WflowModelGroups;
 import com.wflow.bean.entity.WflowModels;
 import com.wflow.bean.vo.ModelGroupVo;
 import com.wflow.bean.vo.WflowModelDetailVo;
+import com.wflow.bean.vo.remote.req.FlowModelGroupRequest;
 import com.wflow.exception.BusinessException;
 import com.wflow.mapper.WflowModelGroupsMapper;
 import com.wflow.mapper.WflowModelsMapper;
@@ -20,6 +22,7 @@ import com.wflow.workflow.service.FormService;
 import com.wflow.workflow.service.ProcessModelService;
 import com.wflow.workflow.service.ProcessNodeCacheService;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -62,6 +65,18 @@ public class FormGroupServiceImpl implements ModelGroupService {
         List<ModelGroupVo.Form> models = Objects.nonNull(userId) ?
                 orgRepositoryService.getModelsByPerm(userId) :
                 modelsMapper.getSysModels();
+        final FlowProcessContext flowProcessContext = FlowProcessContext.getFlowProcessContext();
+        if (flowProcessContext != null && flowProcessContext.getFieldTag() == Boolean.TRUE) {
+            models.clear();
+            final List<WflowModels> wflowModels = modelsMapper.selectList(new LambdaQueryWrapper<WflowModels>()
+                    .notIn(WflowModels::getGroupId, CollectionUtil.newArrayList(0))
+                    .orderByAsc(WflowModels::getSort));
+            for (WflowModels model : wflowModels) {
+                final ModelGroupVo.Form form = new ModelGroupVo.Form();
+                BeanUtils.copyProperties(model, form);
+                models.add(form);
+            }
+        }
         models.forEach(m -> {
             List<ModelGroupVo.Form> forms = groupMap.get(m.getGroupId());
             if (Objects.isNull(forms)){
@@ -72,6 +87,9 @@ public class FormGroupServiceImpl implements ModelGroupService {
         });
         groupsMapper.selectList(new LambdaQueryWrapper<WflowModelGroups>()
                 .orderByAsc(WflowModelGroups::getSort)).forEach(group -> {
+            if (flowProcessContext != null && flowProcessContext.getFieldTag() == Boolean.TRUE &&  CollectionUtil.newArrayList(0L).contains(group.getGroupId())) {
+                return;
+            }
             ModelGroupVo modelGroupVo = ModelGroupVo.builder()
                     .id(group.getGroupId())
                     .name(group.getGroupName())
@@ -79,6 +97,22 @@ public class FormGroupServiceImpl implements ModelGroupService {
                     .build();
             modelGroupVos.add(modelGroupVo);
             modelGroupVo.setItems(groupMap.getOrDefault(group.getGroupId(), Collections.emptyList()));
+        });
+        return modelGroupVos;
+    }
+
+    @Override
+    public List<ModelGroupVo> getGroupModels(String modelName) {
+        List<ModelGroupVo> modelGroupVos = new LinkedList<>();
+        //一次性查出所有表单，再分组归类
+        groupsMapper.selectList(new LambdaQueryWrapper<WflowModelGroups>()
+                .orderByAsc(WflowModelGroups::getSort)).forEach(group -> {
+            ModelGroupVo modelGroupVo = ModelGroupVo.builder()
+                    .id(group.getGroupId())
+                    .name(group.getGroupName())
+                    .items(new LinkedList<>())
+                    .build();
+            modelGroupVos.add(modelGroupVo);
         });
         return modelGroupVos;
     }
@@ -123,7 +157,26 @@ public class FormGroupServiceImpl implements ModelGroupService {
 
     @Override
     public void createModelGroup(String name) {
-        groupsMapper.insert(WflowModelGroups.builder().sort(0).groupName(name).updated(new Date()).build());
+        groupsMapper.insert(
+                WflowModelGroups.builder()
+                        .groupName(name)
+                        .created(new Date())
+                        .status(1)
+                        .sort(0)
+                        .createUser("")
+                        .updated(new Date()).build());
+    }
+
+    @Override
+    public void createModelGroup(FlowModelGroupRequest request) {
+        groupsMapper.insert(
+                WflowModelGroups.builder()
+                        .groupName(request.getGroupName())
+                        .created(new Date())
+                        .status(1)
+                        .sort(request.getSort())
+                        .createUser(request.getCreateUser())
+                        .updated(new Date()).build());
     }
 
     @Override
@@ -140,6 +193,23 @@ public class FormGroupServiceImpl implements ModelGroupService {
                 modelsMapper.update(WflowModels.builder().groupId(1L).build(),
                         new LambdaQueryWrapper<WflowModels>()
                                 .in(WflowModels::getFormId, collect));
+            }
+            groupsMapper.deleteById(id);
+        } else {
+            throw new BusinessException("系统默认分组不允许删除");
+        }
+    }
+    @Override
+    @Transactional
+    public void deleteByGroupId(Long id) {
+        //先转移，再删除分组
+        if (Objects.nonNull(id) && id > 1) {
+            Set<String> collect = modelsMapper.selectList(new LambdaQueryWrapper<WflowModels>()
+                            .select(WflowModels::getFormId)
+                            .eq(WflowModels::getGroupId, id)).stream()
+                    .map(WflowModels::getFormId).collect(Collectors.toSet());
+            if (CollectionUtil.isNotEmpty(collect)) {
+                throw new BusinessException("存在模型分组不允许删除");
             }
             groupsMapper.deleteById(id);
         } else {
